+++ packages/qwik-router/src/middleware/request-handler/response-page.ts
@@ -1,7 +1,5 @@
 import { Q_ROUTE } from '../../runtime/src/constants';
 import {
-  getRequestLoaders,
-  getRequestLoaderSerializationStrategyMap,
   getRequestMode,
   getRequestRoute,
   RequestEvHttpStatusMessage,
@@ -22,8 +20,6 @@ const responsePageDeps = {
   RequestEvSharedActionId,
   RequestEvSharedNonce,
   RequestRouteName,
-  getRequestLoaders,
-  getRequestLoaderSerializationStrategyMap,
   getRequestMode,
   getRequestRoute,
 };