+++ packages/qwik-router/src/runtime/src/index.ts
@@ -23,6 +23,7 @@ export type {
   DocumentStyle,
   FailReturn,
   HttpStatus as HttpErrorProps,
+  InternalRequest,
   JSONObject,
   JSONValue,
   Loader,
@@ -76,8 +77,6 @@ export {
   globalActionQrl,
   routeAction$,
   routeActionQrl,
-  routeLoader$,
-  routeLoaderQrl,
   server$,
   serverQrl,
   valibot$,
@@ -87,6 +86,7 @@ export {
   zod$,
   zodQrl,
 } from './server-functions';
+export { routeLoader$, routeLoaderQrl } from './route-loaders';
 export { ServiceWorkerRegister } from './sw-component';
 export {
   useContent,